+++ commands/proxy.go
@@ -8,16 +8,26 @@ import (
 
 // Raw proxy to docker or docker-compose.
 func Proxy() {
+	var args []string
+	var secondArg string
+
 	application := "docker"
-	args := os.Args[1:]
-	secondArg := os.Args[1]
+
+	if len(os.Args) > 1 {
+		args = os.Args[1:]
+		secondArg = os.Args[1]
+	}
 
 	if secondArg == "c" {
 		application = "docker-compose"
-		args = os.Args[2:]
+		if len(os.Args) > 2 {
+			args = os.Args[2:]
+		} else {
+			args = nil
+		}
 	}
 
-	fmt.Printf("Calling docker with %v\n", args)
+	fmt.Printf("Calling %s with %v\n", application, args)
 	cmd := exec.Command(application, args...)
 	cmd.Stderr = os.Stderr
 	cmd.Stdout = os.Stdout