+++ CHANGELOG.md
@@ -5,7 +5,9 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [Unreleased] (0.1.0)
+## Fixed
+- Index out of range error when calling `dkr` or `dkr c` without any more arguments.
 
 ## [v0.0.1](https://github.com/ohdkr/dkr/releases/tag/v0.0.1)
 ### Added